[Fix] print 32bits address correctly.

+++ library/src/main/java/me/weishu/epic/art/entry/Entry.java
@@ -147,7 +147,7 @@ public class Entry {
         Logger.d(TAG, "r3:" + Debug.hexdump(r3, 0));
 
         final int sourceMethod = ByteBuffer.wrap(EpicNative.get(struct + 12, 4)).order(ByteOrder.LITTLE_ENDIAN).getInt();
-        Logger.i(TAG, "sourceMethod:" + Long.toHexString(sourceMethod));
+        Logger.i(TAG, "sourceMethod:" + Integer.toHexString(sourceMethod));
 
         Epic.MethodInfo originMethodInfo = Epic.getMethodInfo(sourceMethod);
         Logger.i(TAG, "originMethodInfo :" + originMethodInfo);
